perf: speedup boolean filter
